pyproject.toml
[project]
name = "zleditor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "litellm",
    "pydantic-settings",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
